test: DCI 패턴을 적용한 CouponServiceTest 적용
